requirements.txt
pytest
pytest-html
pyyaml
oracledb
pandas